refactor(tasks): move locomotion DR config [skip ci]
