+++ src/unilab/tasks/locomotion/common/__init__.py
@@ -7,9 +7,11 @@ from .commands import (
     sample_velocity_commands,
     zero_small_xy_commands,
 )
+from .domain_rand import DomainRandConfig
 
 __all__ = [
     "Commands",
+    "DomainRandConfig",
     "apply_heading_yaw_feedback",
     "sample_heading_commands",
     "sample_velocity_commands",
