+++ src/unilab/tasks/locomotion/common/domain_rand.py
@@ -1,3 +1,5 @@
+"""Shared domain-randomization config for locomotion tasks."""
+
 from __future__ import annotations
 
 from dataclasses import dataclass, field
